style(indexing): rename verify_manifest_format.py
