+++ gen3/tools/indexing/validate_manifest_format.py
