+++ tests/validate_manifest_format/manifests/manifest_with_custom_url_protocols.tsv
