+++ tests/validate_manifest_format/manifests/manifest_with_invalid_authz_resources.tsv
