+++ tests/validate_manifest_format/manifests/manifest_with_many_types_of_errors.tsv
