+++ tests/validate_manifest_format/manifests/manifest_with_missing_size_column.tsv
