+++ tests/validate_manifest_format/manifests/manifest_with_no_errors.tsv
